Clean up meeting recording tab components
